requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
idna==2.10
instaloader==4.7.4
joblib==1.0.1
kiwisolver==1.3.1
libretranslatepy==2.1.1
lxml==4.6.3
matplotlib==3.4.2
memory-profiler==0.58.0
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
psutil==5.8.0
pyjokes==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
threadpoolctl==2.2.0
translate==3.6.1
urllib3==1.26.4
